% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tus-package.R
\docType{package}
\name{tus-package}
\alias{tus}
\alias{tus-package}
\title{tus: TUS Protocol R client}
\description{
TUS Protocol client based on the Python client and its implementation from 'tator-r' repository. Unlike tator-r this package is only the TUS client itself, update of httr to httr2, and some minor changes and new quality of life functions.
}
\author{
\strong{Maintainer}: Tomas Martinovic \email{tomas.martinovic@vsb.cz} (\href{https://orcid.org/0000-0002-0132-1434}{ORCID})

Other contributors:
\itemize{
  \item VSB - Technical University of Ostrava [copyright holder, funder]
}

}
\keyword{internal}
