% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tus_request.R
\docType{class}
\name{TusBaseRequest}
\alias{TusBaseRequest}
\title{TusBaseRequest}
\description{
TusBaseRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{The request URL}

\item{\code{response_headers}}{Response headers}

\item{\code{status_code}}{HTTP response status code}

\item{\code{response_content}}{Response content}

\item{\code{file}}{File to upload}

\item{\code{request_headers}}{Request headers}

\item{\code{content_length}}{Content length}

\item{\code{upload_checksum}}{Upload checksum}

\item{\code{checksum_algorithm}}{Checksum algorithm}

\item{\code{checksum_algorithm_name}}{Checksum algorithm name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TusBaseRequest-new}{\code{TusBaseRequest$new()}}
\item \href{#method-TusBaseRequest-add_checksum}{\code{TusBaseRequest$add_checksum()}}
\item \href{#method-TusBaseRequest-clone}{\code{TusBaseRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseRequest-new}{}}}
\subsection{Method \code{new()}}{
initialization method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseRequest$new(uploader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uploader}}{The TusUploader instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseRequest-add_checksum"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseRequest-add_checksum}{}}}
\subsection{Method \code{add_checksum()}}{
add_checksum Add the upload checksum header for the chunk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseRequest$add_checksum(chunk)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chunk}}{The chunk to digest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
