% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tus_uploader.R
\docType{class}
\name{TusBaseUploader}
\alias{TusBaseUploader}
\title{TusBaseUploader}
\description{
TusBaseUploader Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file_path}}{File path}

\item{\code{file_stream}}{File stream}

\item{\code{url}}{URL}

\item{\code{client}}{TusClient}

\item{\code{chunk_size}}{Chunk size}

\item{\code{metadata}}{Metadata}

\item{\code{retries}}{Retry limit}

\item{\code{retried}}{Boolean if the request was retried}

\item{\code{retry_delay}}{Retry delay}

\item{\code{store_url}}{Store URL}

\item{\code{url_storage}}{URL storage}

\item{\code{fingerprinter}}{Fingerprinter}

\item{\code{upload_checksum}}{Upload checksum}

\item{\code{default_headers}}{Default headers}

\item{\code{offset}}{Offset}

\item{\code{stop_at}}{Stop at}

\item{\code{request}}{Request}

\item{\code{checksum_algorithm_name}}{Checksum algorithm name}

\item{\code{checksum_algorithm}}{Checksum algorithm}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TusBaseUploader-new}{\code{TusBaseUploader$new()}}
\item \href{#method-TusBaseUploader-get_headers}{\code{TusBaseUploader$get_headers()}}
\item \href{#method-TusBaseUploader-get_url_creation_headers}{\code{TusBaseUploader$get_url_creation_headers()}}
\item \href{#method-TusBaseUploader-get_checksum_algorithm}{\code{TusBaseUploader$get_checksum_algorithm()}}
\item \href{#method-TusBaseUploader-get_checksum_algorithm_name}{\code{TusBaseUploader$get_checksum_algorithm_name()}}
\item \href{#method-TusBaseUploader-get_offset}{\code{TusBaseUploader$get_offset()}}
\item \href{#method-TusBaseUploader-encode_metadata}{\code{TusBaseUploader$encode_metadata()}}
\item \href{#method-TusBaseUploader-init_url_and_offset}{\code{TusBaseUploader$init_url_and_offset()}}
\item \href{#method-TusBaseUploader-set_url}{\code{TusBaseUploader$set_url()}}
\item \href{#method-TusBaseUploader-get_request_length}{\code{TusBaseUploader$get_request_length()}}
\item \href{#method-TusBaseUploader-get_file_stream}{\code{TusBaseUploader$get_file_stream()}}
\item \href{#method-TusBaseUploader-get_file_size}{\code{TusBaseUploader$get_file_size()}}
\item \href{#method-TusBaseUploader-clone}{\code{TusBaseUploader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-new"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-new}{}}}
\subsection{Method \code{new()}}{
initialization method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$new(
  file_path = NULL,
  file_stream = NULL,
  url = NULL,
  client = NULL,
  chunk_size = NULL,
  metadata = NULL,
  retries = 0,
  retry_delay = 30,
  store_url = FALSE,
  url_storage = NULL,
  fingerprinter = NULL,
  upload_checksum = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{File path}

\item{\code{file_stream}}{File stream}

\item{\code{url}}{URL}

\item{\code{client}}{TusClient}

\item{\code{chunk_size}}{Chunk size}

\item{\code{metadata}}{Metadata}

\item{\code{retries}}{Retry limit}

\item{\code{retry_delay}}{Retry delay}

\item{\code{store_url}}{Store URL}

\item{\code{url_storage}}{URL storage}

\item{\code{fingerprinter}}{Fingerprinter}

\item{\code{upload_checksum}}{Upload checksum}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-get_headers"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-get_headers}{}}}
\subsection{Method \code{get_headers()}}{
Get headers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$get_headers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-get_url_creation_headers"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-get_url_creation_headers}{}}}
\subsection{Method \code{get_url_creation_headers()}}{
Get URL creation headers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$get_url_creation_headers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-get_checksum_algorithm"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-get_checksum_algorithm}{}}}
\subsection{Method \code{get_checksum_algorithm()}}{
Get checksum algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$get_checksum_algorithm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-get_checksum_algorithm_name"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-get_checksum_algorithm_name}{}}}
\subsection{Method \code{get_checksum_algorithm_name()}}{
Get checksum algorithm name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$get_checksum_algorithm_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-get_offset"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-get_offset}{}}}
\subsection{Method \code{get_offset()}}{
Get tus upload offset header
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$get_offset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-encode_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-encode_metadata}{}}}
\subsection{Method \code{encode_metadata()}}{
Encode metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$encode_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-init_url_and_offset"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-init_url_and_offset}{}}}
\subsection{Method \code{init_url_and_offset()}}{
Init URL and offset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$init_url_and_offset(url = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{URL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-set_url"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-set_url}{}}}
\subsection{Method \code{set_url()}}{
Set URL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$set_url(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{URL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-get_request_length"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-get_request_length}{}}}
\subsection{Method \code{get_request_length()}}{
Get request length
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$get_request_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-get_file_stream"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-get_file_stream}{}}}
\subsection{Method \code{get_file_stream()}}{
Get file stream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$get_file_stream()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-get_file_size"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-get_file_size}{}}}
\subsection{Method \code{get_file_size()}}{
Get file size
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$get_file_size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusBaseUploader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TusBaseUploader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
