% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tus_client.R
\docType{class}
\name{TusClient}
\alias{TusClient}
\title{TusClient}
\description{
TusClient Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{URL}

\item{\code{headers}}{HTTP headers}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TusClient-new}{\code{TusClient$new()}}
\item \href{#method-TusClient-set_headers}{\code{TusClient$set_headers()}}
\item \href{#method-TusClient-uploader}{\code{TusClient$uploader()}}
\item \href{#method-TusClient-clone}{\code{TusClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-TusClient-new}{}}}
\subsection{Method \code{new()}}{
initialization method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusClient$new(url = NULL, headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{URL}

\item{\code{headers}}{Headers}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusClient-set_headers"></a>}}
\if{latex}{\out{\hypertarget{method-TusClient-set_headers}{}}}
\subsection{Method \code{set_headers()}}{
Set headers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusClient$set_headers(headers = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{headers}}{Headers to set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusClient-uploader"></a>}}
\if{latex}{\out{\hypertarget{method-TusClient-uploader}{}}}
\subsection{Method \code{uploader()}}{
Get the uploader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusClient$uploader(
  file_path = NULL,
  file_stream = NULL,
  url = NULL,
  client = NULL,
  chunk_size = NULL,
  metadata = NULL,
  retries = 10,
  retry_delay = 15,
  store_url = FALSE,
  url_storage = NULL,
  fingerprinter = NULL,
  upload_checksum = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{File path}

\item{\code{file_stream}}{File stream}

\item{\code{url}}{URL}

\item{\code{client}}{TusClient}

\item{\code{chunk_size}}{Chunk size}

\item{\code{metadata}}{Metadata}

\item{\code{retries}}{Retry limit}

\item{\code{retry_delay}}{Retry delay}

\item{\code{store_url}}{Store URL}

\item{\code{url_storage}}{URL storage}

\item{\code{fingerprinter}}{Fingerprinter}

\item{\code{upload_checksum}}{Upload checksum}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TusClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
