% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tus_uploader.R
\docType{class}
\name{TusUploader}
\alias{TusUploader}
\title{TusUploader}
\description{
TusUploader Class
}
\section{Super class}{
\code{tus::TusBaseUploader} -> \code{TusUploader}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{The TusClient instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TusUploader-upload_chunk}{\code{TusUploader$upload_chunk()}}
\item \href{#method-TusUploader-create_url}{\code{TusUploader$create_url()}}
\item \href{#method-TusUploader-do_request}{\code{TusUploader$do_request()}}
\item \href{#method-TusUploader-retry}{\code{TusUploader$retry()}}
\item \href{#method-TusUploader-clone}{\code{TusUploader$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="encode_metadata"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-encode_metadata'><code>tus::TusBaseUploader$encode_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="get_checksum_algorithm"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-get_checksum_algorithm'><code>tus::TusBaseUploader$get_checksum_algorithm()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="get_checksum_algorithm_name"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-get_checksum_algorithm_name'><code>tus::TusBaseUploader$get_checksum_algorithm_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="get_file_size"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-get_file_size'><code>tus::TusBaseUploader$get_file_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="get_file_stream"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-get_file_stream'><code>tus::TusBaseUploader$get_file_stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="get_headers"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-get_headers'><code>tus::TusBaseUploader$get_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="get_offset"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-get_offset'><code>tus::TusBaseUploader$get_offset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="get_request_length"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-get_request_length'><code>tus::TusBaseUploader$get_request_length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="get_url_creation_headers"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-get_url_creation_headers'><code>tus::TusBaseUploader$get_url_creation_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="init_url_and_offset"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-init_url_and_offset'><code>tus::TusBaseUploader$init_url_and_offset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="initialize"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-initialize'><code>tus::TusBaseUploader$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tus" data-topic="TusBaseUploader" data-id="set_url"><a href='../../tus/html/TusBaseUploader.html#method-TusBaseUploader-set_url'><code>tus::TusBaseUploader$set_url()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusUploader-upload_chunk"></a>}}
\if{latex}{\out{\hypertarget{method-TusUploader-upload_chunk}{}}}
\subsection{Method \code{upload_chunk()}}{
Upload a chunk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$upload_chunk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusUploader-create_url"></a>}}
\if{latex}{\out{\hypertarget{method-TusUploader-create_url}{}}}
\subsection{Method \code{create_url()}}{
Create the upload URL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$create_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusUploader-do_request"></a>}}
\if{latex}{\out{\hypertarget{method-TusUploader-do_request}{}}}
\subsection{Method \code{do_request()}}{
Execute the request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$do_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusUploader-retry"></a>}}
\if{latex}{\out{\hypertarget{method-TusUploader-retry}{}}}
\subsection{Method \code{retry()}}{
Retry the upload
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$retry()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TusUploader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TusUploader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
